Feature/sites2 (#31)
(1) Refactor of the power-apps library
(2) Addition of site to the API
(3) Enhancement of documentation

+++ packages/api/src/handlers/application/delete-application.js
@@ -1,8 +1,20 @@
 import { models } from '@defra/wls-database-model'
 import { clearCaches } from './application-cache.js'
 
-export default async (context, req, h) => {
+export default async (context, _req, h) => {
   const { userId, applicationId } = context.request.params
+
+  // Check there are no application sites owned by this application
+  const applicationSites = await models.applicationSites.findAll({
+    where: {
+      userId, applicationId
+    }
+  })
+
+  if (applicationSites.length) {
+    return h.response().code(409)
+  }
+
   await clearCaches(userId, applicationId)
   const count = await models.applications.destroy({
     where: {
